refactor(cli): split skill removal helpers
